Test.Tasty.Hedgehog

Description

This package lets you test Hedgehog properties with tasty.

Typical usage would look like this:

testGroup "tasty-hedgehog tests" [ 
   testProperty "reverse involutive" prop_reverse_involutive
 , testProperty "sort idempotent"    prop_sort_idempotent
 ]

Documentation

testProperty :: TestName -> Property -> TestTree

Create a Test from a Hedgehog property

Options you can pass in via tasty

newtype HedgehogReplay

The replay token to use for replaying a previous test run

Constructors

HedgehogReplay (Maybe (Size, Seed))

newtype HedgehogShowReplay

If a test case fails, show a replay token for replaying tests

Constructors

HedgehogShowReplay Bool

newtype HedgehogVerbose

Show the generated Hedgehog test cases

Constructors

HedgehogVerbose Bool

newtype HedgehogTestLimit

The number of successful test cases required before Hedgehog will pass a test

Constructors

HedgehogTestLimit Int

newtype HedgehogDiscardLimit

The number of discarded cases allowed before Hedgehog will fail a test

Constructors

HedgehogDiscardLimit Int

newtype HedgehogShrinkLimit

The number of shrinks allowed before Hedgehog will fail a test

Constructors

HedgehogShrinkLimit Int

newtype HedgehogShrinkRetries

The number of times to re-run a test during shrinking

Constructors

HedgehogShrinkRetries Int